+++ packages/rust-bindings/src/aux_functions.rs
@@ -3,16 +3,20 @@ use crypto_common::{types::TransactionTime, *};
 use dodis_yampolskiy_prf as prf;
 use pairing::bls12_381::{Bls12, G1};
 use serde_json::{from_str, Value as SerdeValue};
-use std::collections::BTreeMap;
+use std::{cmp::max, collections::BTreeMap, convert::TryInto};
 type ExampleCurve = G1;
 use concordium_contracts_common::{from_bytes, schema, Cursor};
 use hex;
 use serde::{Deserialize as SerdeDeserialize, Serialize as SerdeSerialize};
 use sha2::{Digest, Sha256};
+use key_derivation::{ConcordiumHdWallet, Net};
 
-use anyhow::{anyhow, bail, Result};
-use id::{account_holder::create_unsigned_credential, constants::AttributeKind, types::*};
+use anyhow::{anyhow, bail, ensure, Result};
+use id::{account_holder::{create_unsigned_credential, generate_pio_v1}, constants::{AttributeKind, ArCurve}, types::*, pedersen_commitment::{Value as PedersenValue},};
 use pedersen_scheme::Value;
+use serde_json::{to_string};
+
+use id::secret_sharing::Threshold;
 
 #[derive(SerdeSerialize, SerdeDeserialize)]
 pub struct CredId {
@@ -22,6 +26,89 @@ pub struct CredId {
         deserialize_with = "base16_decode"
     )]
     pub cred_id: ExampleCurve,
+}
+
+
+#[derive(SerdeSerialize, SerdeDeserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct IdRequestInput {
+    ip_info: IpInfo<Bls12>,
+    global_context: GlobalContext<ExampleCurve>,
+    ars_infos: BTreeMap<ArIdentity, ArInfo<ExampleCurve>>,
+    seed: String,
+    net: String,
+    identity_index: u32,
+    ar_threshold: Option<u8>,
+}
+
+pub fn create_id_request_v1_aux(input: IdRequestInput) -> Result<String> {
+    let seed_decoded = hex::decode(&input.seed)?;
+    let seed: [u8; 64] = match seed_decoded.try_into() {
+        Ok(s) => s,
+        Err(_) => bail!("The provided seed {} was not 64 bytes", input.seed),
+    };
+
+    let net = match input.net.as_str() {
+        "Mainnet" => Net::Mainnet,
+        "Testnet" => Net::Testnet,
+        _ => bail!("Unknown net")
+    };
+    let wallet = ConcordiumHdWallet { seed, net };
+
+    let prf_key: prf::SecretKey<ArCurve> = wallet.get_prf_key(input.identity_index)?;
+
+    let id_cred_sec: PedersenValue<ArCurve> =
+        PedersenValue::new(wallet.get_id_cred_sec(input.identity_index)?);
+    let id_cred: IdCredentials<ArCurve> = IdCredentials { id_cred_sec };
+
+    let sig_retrievel_randomness: ps_sig::SigRetrievalRandomness<Bls12> =
+        wallet.get_blinding_randomness(input.identity_index)?;
+
+    let num_of_ars = input.ars_infos.len();
+    let threshold = match input.ar_threshold {
+        Some(threshold) => {
+            ensure!(threshold > 0, "arThreshold must be at least 1.");
+            ensure!(
+                num_of_ars >= usize::from(threshold),
+                "Number of anonymity revokers in arsInfos should be at least arThreshold."
+            );
+            Threshold(threshold)
+        }
+        None => {
+            // arThreshold not specified, use `number of anonymity revokers` - 1 or 1 in the
+            // case of only a single anonymity revoker.
+            ensure!(
+                num_of_ars > 0,
+                "arsInfos should have at least 1 anonymity revoker."
+            );
+            Threshold(max((num_of_ars - 1).try_into().unwrap_or(255), 1))
+        }
+    };
+
+    let chi = CredentialHolderInfo::<ArCurve> { id_cred };
+
+    let aci = AccCredentialInfo {
+        cred_holder_info: chi,
+        prf_key,
+    };
+
+    // Choice of anonymity revokers, all of them in this implementation.
+    let context = IpContext::new(&input.ip_info, &input.ars_infos, &input.global_context);
+
+    let id_use_data = IdObjectUseData {
+        aci,
+        randomness: sig_retrievel_randomness,
+    };
+    let (pio, _) = {
+        match generate_pio_v1(&context, threshold, &id_use_data) {
+            Some(x) => x,
+            None => bail!("Generating the pre-identity object failed."),
+        }
+    };
+
+    let response = json!({ "idObjectRequest": Versioned::new(VERSION_0, pio) });
+
+    Ok(to_string(&response)?)
 }
 
 pub fn generate_unsigned_credential_aux(input: &str) -> Result<String> {
@@ -97,6 +184,7 @@ pub fn generate_unsigned_credential_aux(input: &str) -> Result<String> {
         policy,
         cred_key_info,
         address.as_ref(),
+        &SystemAttributeRandomness {}
     )?;
 
     let response = json!({"unsignedCdi": unsigned_cdi, "randomness": rand});