+++ packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.4.0
+
+### Added
+
+- `createIdentityRequest`, to create identity requests.
+
 ## 2.3.2 2022-7-26
 
 ### Fixed