+++ packages/web/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- `createIdentityRequest`, to create identity requests.
 - `deserializeTransaction` function to deserialize transaction created by `serializeAccountTransactionForSubmission` and `serializeCredentialDeploymentTransactionForSubmission`. (Currently SimpleTransfer, SimpleTransferWithMemo and RegisterData are the only supported account transactions kinds)
 
 ## 0.3.0 2022-7-21